requirements.txt
nltk~=3.9.1
sentence-transformers==2.2.2
pymupdf==1.24.11
PyPDF2==2.12.1
huggingface-hub==0.25.2
transformers==4.45.2
rouge-score==0.1.2

spacy~=3.8.2
numpy~=2.0.2
fitz~=0.0.1.dev2
pinecone~=5.3.1
python-dotenv~=1.0.1
openai~=1.51.2